deploy0.20

+++ backend/recipe_api/models.py
@@ -103,11 +103,7 @@ class RecipeIngredient(models.Model):
         on_delete=models.CASCADE,
         verbose_name="Ингредиент"
     )
-    # amount = models.PositiveSmallIntegerField(
-    #     verbose_name="Количество",
-    #     default=1
-    # )
-    amount = models.FloatField(
+    amount = models.PositiveSmallIntegerField(
         verbose_name="Количество",
         default=1
     )
